workers/audio-split/requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-cloud-tasks==2.16.5
